e2fsprogs: fix unable to find libraries.
The system was looking in /usr/lib64 but not in the usual /usr/lib.

A proper way needs to be made to handle this really.

Perhaps setting --libdir to /usr/lib/x86_64-linux-gnu/ when building glibc and e2fsprogs
would be the ideal thing? Or even a symlink from /usr/lib to
/usr/libARCH.

Honestly I do not care too much about having 32+64 bit support on the
same TOS installation. Although it would be kind of nice.