Commit scripts/mk-usb and refinements.
So far this seems to be an almost there solution.

Sanity improvements would be that setup != recovery, as obviously we
can't reformat ourselves at runtime.

The actual recovery running off a drive would need to have the script
script partitioning and formatting faces of calling the
{mbr,gpt}-install program.

Remaining issues include things like:

  - Making sure dev and boot and images directories are never missing.
  - Full BIOS/UEFI syslinux testing
  - Put a /recovery script on recovery partition.